+++ api/main.go
@@ -62,7 +62,7 @@ func main() {
 		gin.SetMode(gin.ReleaseMode)
 	}
 
-	router := gin.Default()
+	router := common.NewGinEngine(logger, options.Debug)
 	router.Use(common.RequestIdMiddleware)
 	router.Use(common.TokenMiddleware(options.Token))