+++ CHANGELOG.md
@@ -2,9 +2,10 @@
 
 - Improve query binding
 - Add locking for concurrent state (thanks a lot [@abdullah-aghayan](https://github.com/abdullah-aghayan)!)
-- Change logger to Zap
+- Change logger to Zap (include request logger)
 - Added logging to all requests
 - Change unsubscribe type from 2 to 1 (internal type)
+- Fix debug mode never being active
 
 ## v0.3.0 - 2020-05-08